+++ header/pSettings.hpp
@@ -3,10 +3,10 @@
 
 namespace pSettings
 {
-    void setNumRange(int);
-    void setDenomRange(int);
-    int getNumRange();
-    int getDenomRange();
+    void setNumRange(int64_t);
+    void setDenomRange(int64_t);
+    int64_t getNumRange();
+    int64_t getDenomRange();
     bool cheatModeStatus();
     bool verboseModeStatus();
     void page();
