+++ header/polyrand.hpp
@@ -3,5 +3,5 @@
 
 #ifndef POLYNOMIAL_RANDOM
 #define POLYNOMIAL_RANDOM
-std::vector<std::pair<int, int>> randPoly(int);
+std::vector<std::pair<int64_t, int64_t>> randPoly(int64_t);
 #endif
