+++ src/polyrand.cpp
@@ -7,15 +7,15 @@
 
 #include "pSettings.hpp"
 
-std::vector<std::pair<int, int>> randPoly(int degree)
+std::vector<std::pair<int64_t, int64_t>> randPoly(int64_t degree)
 {
-    int num_range = pSettings::getNumRange();
-    int denom_range = pSettings::getDenomRange();
-    std::vector<std::pair<int, int>> result;
-    for (int i = 0; i < degree; i++)
+    int64_t num_range = pSettings::getNumRange();
+    int64_t denom_range = pSettings::getDenomRange();
+    std::vector<std::pair<int64_t, int64_t>> result;
+    for (int64_t i = 0; i < degree; i++)
     {
-        int num = (std::rand() % (num_range * 2)) - num_range;
-        int denom = (std::rand() % denom_range) + 1;
+        int64_t num = (std::rand() % (num_range * 2)) - num_range;
+        int64_t denom = (std::rand() % denom_range) + 1;
         if (std::gcd(num, denom) != 1)
         {
             num /= std::gcd(num, denom);
